requirements.txt
astroid==2.15.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.1
dill==0.3.6
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
platformdirs==3.2.0
pycodestyle==2.10.0
pycparser==2.21
pydocstyle==6.3.0
PyJWT==2.6.0
pylint==2.17.2
requests==2.28.2
snowballstemmer==2.2.0
tomlkit==0.11.7
urllib3==1.26.15
wrapt==1.15.0
